% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\docType{data}
\name{key}
\alias{key}
\title{key data
ver: 0.1.0 in May 2017
author: Dien Giau Bui (aka Richard)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.}
\usage{
key
}
\description{
This is the datasets for `ri_fmatch` function which includes 2 data: name_old and key.
}
\keyword{datasets}
