% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_fmatch.R
\name{ri_fmatch}
\alias{ri_fmatch}
\title{ri_fmatch function}
\usage{
ri_fmatch(name_old, key, method_type = "dl")
}
\description{
ver: 0.1.0
}
\details{
Author: Dien Giau Bui (aka Richard)

In package `rifin`

Description: This function fuzzy-match 2 vector of strings and return the output vector. In this version, I allow only one key.

We need two objects and the method to match:

1. name_old: is the string vector of name (that we need to match)
2. key: a data.frame with at least two variables: name and key (e.g, gvkey or cusip)
3. Method: default is "dl". Other methods:  "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex". Detail: ?stringdist::stringdist

Example:
data("key")
old <- name_old$name_old
match <- ri_fmatch(old, key)
}
