% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mread_csv.R
\name{mread_csv}
\alias{mread_csv}
\title{mread_csv function}
\usage{
mread_csv(location, condition = "")
}
\arguments{
\item{location}{an address of .csv files}

\item{condition}{the condition of the filter}
}
\value{
a data.frame
}
\description{
ver: 0.1.0.

Author: Dien Giau Bui (aka Richard)

This load all .csv file in a folder, filter based on conditions, then rbind (i.e., rows combine) together
}
\examples{
mread_csv("home/", x > 0)

}
